# v7: implement GetValue method on flags