This means that there's a common API for all flags
that can be used to get a flag's value as an interface.
This is technically unnecessary - a caller could use
`Snapshot.GetValue` - but using `Flag.GetValue` means
the you get to use the flag's default and the precalculated
key id that's inside the flag, which makes it more efficient.

This gives a few ns of penalty in the lowrer-cost benchmarks,
but nothing to worry about IMHO - 10ns for a call is still pretty good :)

```
name                                   old time/op    new time/op    delta
Get/one-of/get-and-make-8                 241ns ± 1%     242ns ± 3%     ~     (p=0.841 n=5+5)
Get/one-of/get-only-8                    7.40ns ± 1%    9.31ns ± 0%  +25.86%  (p=0.008 n=5+5)
Get/less-than-with-int/get-and-make-8     204ns ±11%     197ns ± 0%     ~     (p=0.690 n=5+5)
Get/less-than-with-int/get-only-8        7.38ns ± 1%    9.35ns ± 1%  +26.72%  (p=0.008 n=5+5)
Get/with-percentage/get-and-make-8        494ns ± 2%     486ns ± 0%   -1.61%  (p=0.008 n=5+5)
Get/with-percentage/get-only-8           7.36ns ± 0%    9.50ns ± 0%  +29.19%  (p=0.016 n=5+4)
Get/no-rules/get-and-make-8               117ns ± 0%     121ns ± 0%   +2.69%  (p=0.016 n=5+4)
Get/no-rules/get-only-8                  4.33ns ± 0%    6.07ns ± 0%  +40.24%  (p=0.008 n=5+5)
Get/no-user/get-and-make-8               10.6ns ± 0%    12.5ns ± 0%  +18.22%  (p=0.008 n=5+5)
Get/no-user/get-only-8                   7.37ns ± 0%    9.30ns ± 0%  +26.30%  (p=0.016 n=5+4)
NewSnapshot-8                            7.83µs ± 1%    8.96µs ±27%  +14.43%  (p=0.008 n=5+5)

name                                   old alloc/op   new alloc/op   delta
Get/one-of/get-and-make-8                  180B ± 0%      180B ± 0%     ~     (all equal)
Get/one-of/get-only-8                     0.00B          0.00B          ~     (all equal)
Get/less-than-with-int/get-and-make-8      180B ± 0%      180B ± 0%     ~     (all equal)
Get/less-than-with-int/get-only-8         0.00B          0.00B          ~     (all equal)
Get/with-percentage/get-and-make-8         252B ± 0%      252B ± 0%     ~     (all equal)
Get/with-percentage/get-only-8            0.00B          0.00B          ~     (all equal)
Get/no-rules/get-and-make-8                176B ± 0%      176B ± 0%     ~     (all equal)
Get/no-rules/get-only-8                   0.00B          0.00B          ~     (all equal)
Get/no-user/get-and-make-8                0.00B          0.00B          ~     (all equal)
Get/no-user/get-only-8                    0.00B          0.00B          ~     (all equal)
NewSnapshot-8                            5.10kB ± 0%    5.10kB ± 0%     ~     (all equal)

name                                   old allocs/op  new allocs/op  delta
Get/one-of/get-and-make-8                  2.00 ± 0%      2.00 ± 0%     ~     (all equal)
Get/one-of/get-only-8                      0.00           0.00          ~     (all equal)
Get/less-than-with-int/get-and-make-8      2.00 ± 0%      2.00 ± 0%     ~     (all equal)
Get/less-than-with-int/get-only-8          0.00           0.00          ~     (all equal)
Get/with-percentage/get-and-make-8         4.00 ± 0%      4.00 ± 0%     ~     (all equal)
Get/with-percentage/get-only-8             0.00           0.00          ~     (all equal)
Get/no-rules/get-and-make-8                1.00 ± 0%      1.00 ± 0%     ~     (all equal)
Get/no-rules/get-only-8                    0.00           0.00          ~     (all equal)
Get/no-user/get-and-make-8                 0.00           0.00          ~     (all equal)
Get/no-user/get-only-8                     0.00           0.00          ~     (all equal)
NewSnapshot-8                              6.00 ± 0%      6.00 ± 0%     ~     (all equal)
```

### Describe the purpose of your pull request

Provide a clear and concise description of the changes.

### Related issues (only if applicable)

Provide links to issues relating to this pull request

### Requirement checklist (only if applicable)

- [ ] I have covered the applied changes with automated tests.
- [ ] I have executed the full automated test set against my changes.
- [ ] I have validated my changes against all supported platform versions.
- [ ] I have read and accepted the [contribution agreement](https://github.com/configcat/legal/blob/main/contribution-agreement.md).